add delivery type, rename platofrm.py
